Keep track of factories in Control object
It's really hard to send messages to the Protocol objects from outside,
but that's where a lot of the working code is for the IRC client. This
code lets a Control object keep track of the ClientFactories, and lets
the ClientFactories keep track of their Protocol objects. A join message
sent to the Factories gets reflected down the chain. We use deferrals to
make sure it arrives at a Protocol, even when it hasn't yet signed on
when the join method is called.